client connects to server
