+++ Cargo.toml
@@ -5,3 +5,14 @@ edition = "2024"
 
 [dependencies]
 minifb = "0.28.0"
+serde = { version = "1.0", features = ["derive"] }
+bincode = "1.3"
+local-ip-address = "0.6.1"
+
+[[bin]]
+name = "server"
+path = "src/bin/server.rs"
+
+[[bin]]
+name = "client"
+path = "src/bin/client.rs"
